Rename coordinator.py
