+++ .coveragerc
@@ -497,6 +497,7 @@ omit =
     homeassistant/components/hive/water_heater.py
     homeassistant/components/hko/__init__.py
     homeassistant/components/hko/weather.py
+    homeassistant/components/hko/coordinator.py
     homeassistant/components/hlk_sw16/__init__.py
     homeassistant/components/hlk_sw16/switch.py
     homeassistant/components/home_connect/__init__.py
